task #54757 Resource > All resource: API view retro & intern review
