+++ aspnet-core/src/ProjectManagement.Core/Services/ResourceService/ResourceManager.cs
@@ -124,19 +124,19 @@ namespace ProjectManagement.Services.ResourceManager
         public IQueryable<ProjectOfUserDto> QueryWorkingProjectsOfUser(long userId)
         {
             return _workScope.GetAll<ProjectUser>()
-                 .Where(s => s.UserId == userId)
-                 .Where(s => s.Status == ProjectUserStatus.Present)
-                 .Where(s => s.AllocatePercentage > 0)
-                 .Where(s => s.Project.Status != ProjectStatus.Closed)
-                 .OrderBy(s => s.ProjectRole)
-                 .ThenByDescending(s => s.StartTime)
-                 .Select(x => new ProjectOfUserDto
-                 {
-                     ProjectName = x.Project.Name,
-                     PmName = x.Project.PM.FullName,
-                     StartTime = x.StartTime,
-                     IsPool = x.IsPool
-                 });
+                .Where(s => s.UserId == userId)
+                .Where(s => s.Status == ProjectUserStatus.Present)
+                .Where(s => s.AllocatePercentage > 0)
+                .Where(s => s.Project.Status == ProjectStatus.InProgress)
+                .OrderBy(s => s.ProjectRole)
+                .ThenByDescending(s => s.StartTime)
+                .Select(x => new ProjectOfUserDto
+                {
+                    ProjectName = x.Project.Name,
+                    PmName = x.Project.PM.FullName,
+                    StartTime = x.StartTime,
+                    IsPool = x.IsPool
+                });
         }
 
         public List<PMOfUserDto> QueryPMOfUser(long userId)
